lesson5/requirements.txt
allure-pytest==2.13.5
Faker==33.0.0
playwright==1.48.0
pytest-playwright==0.5.2
